services/event/pyproject.toml
[project]
name = "cmdb-event"
version = "0.1.0"
description = "CMDB event Service"
requires-python = ">=3.13"
dependencies = [
    "cmdb-shared",
    "fastapi>=0.115",
    "uvicorn",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "pydantic-settings>=2.0",
    "aiokafka",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/event"]

[tool.uv.sources]
cmdb-shared = { workspace = true }
